+++ test/jest/controls/control.selectMany.test.js
@@ -69,7 +69,7 @@ describe("control.selectMany", () => {
 
     el.focus(); // opening menu by focus
     await h.wait(1);
-    expect(el.$isOpen).toBe(true);
+    expect(el.$isShown).toBe(true);
     expect(el.innerHTML).toMatchInlineSnapshot(
       `"<label for="txt1"><span><span item="" aria-hidden="true">Donny</span><span item="" aria-hidden="true">Leo</span><input placeholder=" " type="text" id="txt1" role="combobox" aria-haspopup="listbox" aria-expanded="true" autocomplete="off" aria-autocomplete="list" class="" aria-describedby="txt2" aria-owns="txt3" aria-controls="txt3"><strong></strong></span><button clear="" tabindex="-1" aria-hidden="true" type="button"></button></label><section aria-live="off" aria-atomic="true" class="wup-hidden" id="txt2">Donny,Leo</section><wup-popup menu="" style="min-width: 100px;"><ul id="txt3" role="listbox" aria-label="Items" tabindex="-1" aria-multiselectable="true"><li role="option" style="display: none;">Donny</li><li role="option">Mikky</li><li role="option" style="display: none;">Leo</li><li role="option">Splinter</li></ul></wup-popup>"`
     );
@@ -87,7 +87,7 @@ describe("control.selectMany", () => {
     );
     await h.userClick(el.$refPopup.querySelectorAll("li")[1]); // select first visible
     await h.wait();
-    expect(el.$isOpen).toBe(true);
+    expect(el.$isShown).toBe(true);
     expect(el.$value).toStrictEqual([10, 30, 20]);
     expect(el.$refInput.value).toBe(""); // input is cleared after selection
     expect(el.$refPopup.innerHTML).toMatchInlineSnapshot(
@@ -109,7 +109,7 @@ describe("control.selectMany", () => {
     await h.wait();
     expect(el.$value).toStrictEqual([10, 30, 20, 40]); // because selected Splinter
     expect(el.$refInput.value).toBe("");
-    expect(el.$isOpen).toBe(false); // because no-items anymore
+    expect(el.$isShown).toBe(false); // because no-items anymore
     expect(el.$refPopup.innerHTML).toMatchInlineSnapshot(
       `"<ul id="txt3" role="listbox" aria-label="Items" tabindex="-1" aria-multiselectable="true"><li role="option" style="display: none;">Donny</li><li role="option" aria-selected="false" id="txt4" style="display: none;">Mikky</li><li role="option" style="display: none;">Leo</li><li role="option" style="display: none;" aria-selected="false" id="txt8">Splinter</li><li role="option" aria-disabled="true" aria-selected="false">No Items</li></ul>"`
     );
@@ -117,7 +117,7 @@ describe("control.selectMany", () => {
     // expected no-items
     el.$refInput.dispatchEvent(new KeyboardEvent("keydown", { key: "ArrowDown", bubbles: true, cancelable: true }));
     await h.wait();
-    expect(el.$isOpen).toBe(true);
+    expect(el.$isShown).toBe(true);
     expect(el.$refInput.outerHTML).toMatchInlineSnapshot(
       `"<input placeholder=" " type="text" id="txt1" role="combobox" aria-haspopup="listbox" aria-expanded="true" autocomplete="off" aria-autocomplete="list" class="" aria-owns="txt3" aria-controls="txt3">"`
     );
@@ -127,7 +127,7 @@ describe("control.selectMany", () => {
 
     await h.userClick(el.$refItems[0]); // remove item by click
     await h.wait();
-    expect(el.$isOpen).toBe(true);
+    expect(el.$isShown).toBe(true);
     expect(el.$value).toStrictEqual([30, 20, 40]); // because selected Splinter
     expect(el.$refPopup.innerHTML).toMatchInlineSnapshot(
       `"<ul id="txt3" role="listbox" aria-label="Items" tabindex="-1" aria-multiselectable="true"><li role="option" style="">Donny</li><li role="option" aria-selected="false" id="txt4" style="display: none;">Mikky</li><li role="option" style="display: none;">Leo</li><li role="option" style="display: none;" aria-selected="false" id="txt8">Splinter</li><li role="option" aria-disabled="true" aria-selected="false" style="display: none;">No Items</li></ul>"`
@@ -159,7 +159,7 @@ describe("control.selectMany", () => {
     await h.userTap(item);
     await h.wait();
     expect(el.$isFocused).toBe(true);
-    expect(el.$isOpen).toBe(true); // 1st tap only to focus control
+    expect(el.$isShown).toBe(true); // 1st tap only to focus control
     // item must be not-removed
     expect(el.$refInput.parentElement.innerHTML).toMatchInlineSnapshot(
       `"<span item="" aria-hidden="true">Leo</span><span item="" aria-hidden="true">Mikky</span><span item="" aria-hidden="true">Splinter</span><input placeholder=" " type="text" id="txt1" role="combobox" aria-haspopup="listbox" aria-expanded="true" autocomplete="off" aria-autocomplete="list" class="" aria-owns="txt14" aria-controls="txt14"><strong></strong>"`
@@ -234,7 +234,7 @@ describe("control.selectMany", () => {
     expect(el.$refInput.parentElement.innerHTML).toMatchInlineSnapshot(
       `"<span item="" aria-hidden="true">Leo</span><input placeholder=" " type="text" id="txt1" role="combobox" aria-haspopup="listbox" aria-expanded="false" autocomplete="off" aria-autocomplete="list" class=""><strong></strong>"`
     );
-    expect(el.$isOpen).toBe(false); // because not-open by value-change
+    expect(el.$isShown).toBe(false); // because not-open by value-change
 
     // without animation
     jest.spyOn(window, "matchMedia").mockImplementation(() => ({ matches: true })); // simulate 'prefers-reduced-motion' === true
@@ -250,7 +250,7 @@ describe("control.selectMany", () => {
     expect(el.$refInput.parentElement.innerHTML).toMatchInlineSnapshot(
       `"<input placeholder=" " type="text" id="txt1" role="combobox" aria-haspopup="listbox" aria-expanded="false" autocomplete="off" aria-autocomplete="list" class=""><strong></strong>"`
     );
-    expect(el.$isOpen).toBe(false); // because not-open by value-change
+    expect(el.$isShown).toBe(false); // because not-open by value-change
   });
 
   test("menu: focus behavior", async () => {
@@ -258,7 +258,7 @@ describe("control.selectMany", () => {
     el.$value = [getItems()[0].value];
     el.focus();
     await h.wait();
-    expect(el.$isOpen).toBe(true);
+    expect(el.$isShown).toBe(true);
 
     expect(el.$refPopup.innerHTML).toMatchInlineSnapshot(
       `"<ul id="txt3" role="listbox" aria-label="Items" tabindex="-1" aria-multiselectable="true"><li role="option" style="display: none;">Donny</li><li role="option">Mikky</li><li role="option">Leo</li><li role="option">Splinter</li></ul>"`
@@ -270,7 +270,7 @@ describe("control.selectMany", () => {
     el.dispatchEvent(new KeyboardEvent("keydown", { key: "Enter", bubbles: true, cancelable: true }));
     await h.wait();
     expect(el.$value).toStrictEqual([10, 20]);
-    expect(el.$isOpen).toBe(true);
+    expect(el.$isShown).toBe(true);
     expect(el.querySelector("[focused]")).toBeFalsy();
 
     el.dispatchEvent(new KeyboardEvent("keydown", { key: "ArrowDown", bubbles: true, cancelable: true }));
@@ -286,7 +286,7 @@ describe("control.selectMany", () => {
     el.$value = [getItems()[0].value];
     el.focus();
     await h.wait();
-    expect(el.$isOpen).toBe(true);
+    expect(el.$isShown).toBe(true);
     expect(el.$refPopup.innerHTML).toMatchInlineSnapshot(
       `"<ul id="txt3" role="listbox" aria-label="Items" tabindex="-1" aria-multiselectable="true"><li role="option" aria-selected="true">Donny</li><li role="option">Mikky</li><li role="option">Leo</li><li role="option">Splinter</li></ul>"`
     );
@@ -548,7 +548,7 @@ describe("control.selectMany", () => {
       ]
     `);
     expect(el.$refInput.value).toBe("");
-    expect(el.$isOpen).toBe(true);
+    expect(el.$isShown).toBe(true);
     expect(onChange).toBeCalledTimes(1);
 
     expect(handledKeydown("Enter")).toBe(true);